First step in refactoring blob: containers
