+++ storage/blob.go
@@ -13,45 +13,6 @@ import (
 	"time"
 )
 
-// BlobServiceClient contains operations for Microsoft Azure Blob Storage
-// Service.
-type BlobServiceClient struct {
-	client Client
-	auth   authentication
-}
-
-// A Container is an entry in ContainerListResponse.
-type Container struct {
-	Name       string              `xml:"Name"`
-	Properties ContainerProperties `xml:"Properties"`
-	// TODO (ahmetalpbalkan) Metadata
-}
-
-// ContainerProperties contains various properties of a container returned from
-// various endpoints like ListContainers.
-type ContainerProperties struct {
-	LastModified  string `xml:"Last-Modified"`
-	Etag          string `xml:"Etag"`
-	LeaseStatus   string `xml:"LeaseStatus"`
-	LeaseState    string `xml:"LeaseState"`
-	LeaseDuration string `xml:"LeaseDuration"`
-	// TODO (ahmetalpbalkan) remaining fields
-}
-
-// ContainerListResponse contains the response fields from
-// ListContainers call.
-//
-// See https://msdn.microsoft.com/en-us/library/azure/dd179352.aspx
-type ContainerListResponse struct {
-	XMLName    xml.Name    `xml:"EnumerationResults"`
-	Xmlns      string      `xml:"xmlns,attr"`
-	Prefix     string      `xml:"Prefix"`
-	Marker     string      `xml:"Marker"`
-	NextMarker string      `xml:"NextMarker"`
-	MaxResults int64       `xml:"MaxResults"`
-	Containers []Container `xml:"Containers>Container"`
-}
-
 // A Blob is an entry in BlobListResponse.
 type Blob struct {
 	Name       string         `xml:"Name"`
@@ -136,101 +97,6 @@ type BlobHeaders struct {
 	CacheControl    string `header:"x-ms-blob-cache-control"`
 }
 
-// BlobListResponse contains the response fields from ListBlobs call.
-//
-// See https://msdn.microsoft.com/en-us/library/azure/dd135734.aspx
-type BlobListResponse struct {
-	XMLName    xml.Name `xml:"EnumerationResults"`
-	Xmlns      string   `xml:"xmlns,attr"`
-	Prefix     string   `xml:"Prefix"`
-	Marker     string   `xml:"Marker"`
-	NextMarker string   `xml:"NextMarker"`
-	MaxResults int64    `xml:"MaxResults"`
-	Blobs      []Blob   `xml:"Blobs>Blob"`
-
-	// BlobPrefix is used to traverse blobs as if it were a file system.
-	// It is returned if ListBlobsParameters.Delimiter is specified.
-	// The list here can be thought of as "folders" that may contain
-	// other folders or blobs.
-	BlobPrefixes []string `xml:"Blobs>BlobPrefix>Name"`
-
-	// Delimiter is used to traverse blobs as if it were a file system.
-	// It is returned if ListBlobsParameters.Delimiter is specified.
-	Delimiter string `xml:"Delimiter"`
-}
-
-// ListContainersParameters defines the set of customizable parameters to make a
-// List Containers call.
-//
-// See https://msdn.microsoft.com/en-us/library/azure/dd179352.aspx
-type ListContainersParameters struct {
-	Prefix     string
-	Marker     string
-	Include    string
-	MaxResults uint
-	Timeout    uint
-}
-
-func (p ListContainersParameters) getParameters() url.Values {
-	out := url.Values{}
-
-	if p.Prefix != "" {
-		out.Set("prefix", p.Prefix)
-	}
-	if p.Marker != "" {
-		out.Set("marker", p.Marker)
-	}
-	if p.Include != "" {
-		out.Set("include", p.Include)
-	}
-	if p.MaxResults != 0 {
-		out.Set("maxresults", fmt.Sprintf("%v", p.MaxResults))
-	}
-	if p.Timeout != 0 {
-		out.Set("timeout", fmt.Sprintf("%v", p.Timeout))
-	}
-
-	return out
-}
-
-// ListBlobsParameters defines the set of customizable
-// parameters to make a List Blobs call.
-//
-// See https://msdn.microsoft.com/en-us/library/azure/dd135734.aspx
-type ListBlobsParameters struct {
-	Prefix     string
-	Delimiter  string
-	Marker     string
-	Include    string
-	MaxResults uint
-	Timeout    uint
-}
-
-func (p ListBlobsParameters) getParameters() url.Values {
-	out := url.Values{}
-
-	if p.Prefix != "" {
-		out.Set("prefix", p.Prefix)
-	}
-	if p.Delimiter != "" {
-		out.Set("delimiter", p.Delimiter)
-	}
-	if p.Marker != "" {
-		out.Set("marker", p.Marker)
-	}
-	if p.Include != "" {
-		out.Set("include", p.Include)
-	}
-	if p.MaxResults != 0 {
-		out.Set("maxresults", fmt.Sprintf("%v", p.MaxResults))
-	}
-	if p.Timeout != 0 {
-		out.Set("timeout", fmt.Sprintf("%v", p.Timeout))
-	}
-
-	return out
-}
-
 // BlobType defines the type of the Azure Blob.
 type BlobType string
 
@@ -287,41 +153,6 @@ const (
 	BlockListTypeAll         BlockListType = "all"
 	BlockListTypeCommitted   BlockListType = "committed"
 	BlockListTypeUncommitted BlockListType = "uncommitted"
-)
-
-// ContainerAccessType defines the access level to the container from a public
-// request.
-//
-// See https://msdn.microsoft.com/en-us/library/azure/dd179468.aspx and "x-ms-
-// blob-public-access" header.
-type ContainerAccessType string
-
-// Access options for containers
-const (
-	ContainerAccessTypePrivate   ContainerAccessType = ""
-	ContainerAccessTypeBlob      ContainerAccessType = "blob"
-	ContainerAccessTypeContainer ContainerAccessType = "container"
-)
-
-// ContainerAccessPolicyDetails are used for SETTING container policies
-type ContainerAccessPolicyDetails struct {
-	ID         string
-	StartTime  time.Time
-	ExpiryTime time.Time
-	CanRead    bool
-	CanWrite   bool
-	CanDelete  bool
-}
-
-// ContainerPermissions is used when setting permissions and Access Policies for containers.
-type ContainerPermissions struct {
-	AccessType     ContainerAccessType
-	AccessPolicies []ContainerAccessPolicyDetails
-}
-
-// ContainerAccessHeader references header used when setting/getting container ACL
-const (
-	ContainerAccessHeader string = "x-ms-blob-public-access"
 )
 
 // Maximum sizes (per REST API) for various concepts
@@ -386,233 +217,6 @@ var (
 	errBlobCopyAborted    = errors.New("storage: blob copy is aborted")
 	errBlobCopyIDMismatch = errors.New("storage: blob copy id is a mismatch")
 )
-
-// ListContainers returns the list of containers in a storage account along with
-// pagination token and other response details.
-//
-// See https://msdn.microsoft.com/en-us/library/azure/dd179352.aspx
-func (b BlobServiceClient) ListContainers(params ListContainersParameters) (ContainerListResponse, error) {
-	q := mergeParams(params.getParameters(), url.Values{"comp": {"list"}})
-	uri := b.client.getEndpoint(blobServiceName, "", q)
-	headers := b.client.getStandardHeaders()
-
-	var out ContainerListResponse
-	resp, err := b.client.exec(http.MethodGet, uri, headers, nil, b.auth)
-	if err != nil {
-		return out, err
-	}
-	defer resp.body.Close()
-
-	err = xmlUnmarshal(resp.body, &out)
-	return out, err
-}
-
-// CreateContainer creates a blob container within the storage account
-// with given name and access level. Returns error if container already exists.
-//
-// See https://msdn.microsoft.com/en-us/library/azure/dd179468.aspx
-func (b BlobServiceClient) CreateContainer(name string, access ContainerAccessType) error {
-	resp, err := b.createContainer(name, access)
-	if err != nil {
-		return err
-	}
-	defer resp.body.Close()
-	return checkRespCode(resp.statusCode, []int{http.StatusCreated})
-}
-
-// CreateContainerIfNotExists creates a blob container if it does not exist. Returns
-// true if container is newly created or false if container already exists.
-func (b BlobServiceClient) CreateContainerIfNotExists(name string, access ContainerAccessType) (bool, error) {
-	resp, err := b.createContainer(name, access)
-	if resp != nil {
-		defer resp.body.Close()
-		if resp.statusCode == http.StatusCreated || resp.statusCode == http.StatusConflict {
-			return resp.statusCode == http.StatusCreated, nil
-		}
-	}
-	return false, err
-}
-
-func (b BlobServiceClient) createContainer(name string, access ContainerAccessType) (*storageResponse, error) {
-	uri := b.client.getEndpoint(blobServiceName, pathForContainer(name), url.Values{"restype": {"container"}})
-
-	headers := b.client.getStandardHeaders()
-	if access != "" {
-		headers[ContainerAccessHeader] = string(access)
-	}
-	return b.client.exec(http.MethodPut, uri, headers, nil, b.auth)
-}
-
-// ContainerExists returns true if a container with given name exists
-// on the storage account, otherwise returns false.
-func (b BlobServiceClient) ContainerExists(name string) (bool, error) {
-	uri := b.client.getEndpoint(blobServiceName, pathForContainer(name), url.Values{"restype": {"container"}})
-	headers := b.client.getStandardHeaders()
-
-	resp, err := b.client.exec(http.MethodHead, uri, headers, nil, b.auth)
-	if resp != nil {
-		defer resp.body.Close()
-		if resp.statusCode == http.StatusOK || resp.statusCode == http.StatusNotFound {
-			return resp.statusCode == http.StatusOK, nil
-		}
-	}
-	return false, err
-}
-
-// SetContainerPermissions sets up container permissions as per https://msdn.microsoft.com/en-us/library/azure/dd179391.aspx
-func (b BlobServiceClient) SetContainerPermissions(container string, containerPermissions ContainerPermissions, timeout int, leaseID string) (err error) {
-	params := url.Values{
-		"restype": {"container"},
-		"comp":    {"acl"},
-	}
-
-	if timeout > 0 {
-		params.Add("timeout", strconv.Itoa(timeout))
-	}
-
-	uri := b.client.getEndpoint(blobServiceName, pathForContainer(container), params)
-	headers := b.client.getStandardHeaders()
-	if containerPermissions.AccessType != "" {
-		headers[ContainerAccessHeader] = string(containerPermissions.AccessType)
-	}
-
-	if leaseID != "" {
-		headers[headerLeaseID] = leaseID
-	}
-
-	body, length, err := generateContainerACLpayload(containerPermissions.AccessPolicies)
-	headers["Content-Length"] = strconv.Itoa(length)
-	resp, err := b.client.exec(http.MethodPut, uri, headers, body, b.auth)
-
-	if err != nil {
-		return err
-	}
-
-	if resp != nil {
-		defer resp.body.Close()
-
-		if resp.statusCode != http.StatusOK {
-			return errors.New("Unable to set permissions")
-		}
-	}
-	return nil
-}
-
-// GetContainerPermissions gets the container permissions as per https://msdn.microsoft.com/en-us/library/azure/dd179469.aspx
-// If timeout is 0 then it will not be passed to Azure
-// leaseID will only be passed to Azure if populated
-// Returns permissionResponse which is combined permissions and AccessPolicy
-func (b BlobServiceClient) GetContainerPermissions(container string, timeout int, leaseID string) (*ContainerPermissions, error) {
-	params := url.Values{"restype": {"container"},
-		"comp": {"acl"}}
-
-	if timeout > 0 {
-		params.Add("timeout", strconv.Itoa(timeout))
-	}
-
-	uri := b.client.getEndpoint(blobServiceName, pathForContainer(container), params)
-	headers := b.client.getStandardHeaders()
-
-	if leaseID != "" {
-		headers[headerLeaseID] = leaseID
-	}
-
-	resp, err := b.client.exec(http.MethodGet, uri, headers, nil, b.auth)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.body.Close()
-
-	var out AccessPolicy
-	err = xmlUnmarshal(resp.body, &out.SignedIdentifiersList)
-	if err != nil {
-		return nil, err
-	}
-
-	permissionResponse := updateContainerAccessPolicy(out, &resp.headers)
-	return &permissionResponse, nil
-}
-
-func updateContainerAccessPolicy(ap AccessPolicy, headers *http.Header) ContainerPermissions {
-	// containerAccess. Blob, Container, empty
-	containerAccess := headers.Get(http.CanonicalHeaderKey(ContainerAccessHeader))
-
-	var cp ContainerPermissions
-	cp.AccessType = ContainerAccessType(containerAccess)
-	for _, policy := range ap.SignedIdentifiersList.SignedIdentifiers {
-		capd := ContainerAccessPolicyDetails{
-			ID:         policy.ID,
-			StartTime:  policy.AccessPolicy.StartTime,
-			ExpiryTime: policy.AccessPolicy.ExpiryTime,
-		}
-		capd.CanRead = updatePermissions(policy.AccessPolicy.Permission, "r")
-		capd.CanWrite = updatePermissions(policy.AccessPolicy.Permission, "w")
-		capd.CanDelete = updatePermissions(policy.AccessPolicy.Permission, "d")
-
-		cp.AccessPolicies = append(cp.AccessPolicies, capd)
-	}
-
-	return cp
-}
-
-// DeleteContainer deletes the container with given name on the storage
-// account. If the container does not exist returns error.
-//
-// See https://msdn.microsoft.com/en-us/library/azure/dd179408.aspx
-func (b BlobServiceClient) DeleteContainer(name string) error {
-	resp, err := b.deleteContainer(name)
-	if err != nil {
-		return err
-	}
-	defer resp.body.Close()
-	return checkRespCode(resp.statusCode, []int{http.StatusAccepted})
-}
-
-// DeleteContainerIfExists deletes the container with given name on the storage
-// account if it exists. Returns true if container is deleted with this call, or
-// false if the container did not exist at the time of the Delete Container
-// operation.
-//
-// See https://msdn.microsoft.com/en-us/library/azure/dd179408.aspx
-func (b BlobServiceClient) DeleteContainerIfExists(name string) (bool, error) {
-	resp, err := b.deleteContainer(name)
-	if resp != nil {
-		defer resp.body.Close()
-		if resp.statusCode == http.StatusAccepted || resp.statusCode == http.StatusNotFound {
-			return resp.statusCode == http.StatusAccepted, nil
-		}
-	}
-	return false, err
-}
-
-func (b BlobServiceClient) deleteContainer(name string) (*storageResponse, error) {
-	uri := b.client.getEndpoint(blobServiceName, pathForContainer(name), url.Values{"restype": {"container"}})
-
-	headers := b.client.getStandardHeaders()
-	return b.client.exec(http.MethodDelete, uri, headers, nil, b.auth)
-}
-
-// ListBlobs returns an object that contains list of blobs in the container,
-// pagination token and other information in the response of List Blobs call.
-//
-// See https://msdn.microsoft.com/en-us/library/azure/dd135734.aspx
-func (b BlobServiceClient) ListBlobs(container string, params ListBlobsParameters) (BlobListResponse, error) {
-	q := mergeParams(params.getParameters(), url.Values{
-		"restype": {"container"},
-		"comp":    {"list"}})
-	uri := b.client.getEndpoint(blobServiceName, pathForContainer(container), q)
-	headers := b.client.getStandardHeaders()
-
-	var out BlobListResponse
-	resp, err := b.client.exec(http.MethodGet, uri, headers, nil, b.auth)
-	if err != nil {
-		return out, err
-	}
-	defer resp.body.Close()
-
-	err = xmlUnmarshal(resp.body, &out)
-	return out, err
-}
 
 // BlobExists returns true if a blob with given name exists on the specified
 // container of the storage account.
@@ -1402,11 +1006,6 @@ func (b BlobServiceClient) deleteBlob(container, name string, extraHeaders map[s
 	return b.client.exec(http.MethodDelete, uri, headers, nil, b.auth)
 }
 
-// helper method to construct the path to a container given its name
-func pathForContainer(name string) string {
-	return fmt.Sprintf("/%s", name)
-}
-
 // helper method to construct the path to a blob given its container and blob
 // name
 func pathForBlob(container, name string) string {
@@ -1504,36 +1103,4 @@ func blobSASStringToSign(signedVersion, canonicalizedResource, signedExpiry, sig
 	}
 
 	return "", errors.New("storage: not implemented SAS for versions earlier than 2013-08-15")
-}
-
-func generateContainerACLpayload(policies []ContainerAccessPolicyDetails) (io.Reader, int, error) {
-	sil := SignedIdentifiers{
-		SignedIdentifiers: []SignedIdentifier{},
-	}
-	for _, capd := range policies {
-		permission := capd.generateContainerPermissions()
-		signedIdentifier := convertAccessPolicyToXMLStructs(capd.ID, capd.StartTime, capd.ExpiryTime, permission)
-		sil.SignedIdentifiers = append(sil.SignedIdentifiers, signedIdentifier)
-	}
-	return xmlMarshal(sil)
-}
-
-func (capd *ContainerAccessPolicyDetails) generateContainerPermissions() (permissions string) {
-	// generate the permissions string (rwd).
-	// still want the end user API to have bool flags.
-	permissions = ""
-
-	if capd.CanRead {
-		permissions += "r"
-	}
-
-	if capd.CanWrite {
-		permissions += "w"
-	}
-
-	if capd.CanDelete {
-		permissions += "d"
-	}
-
-	return permissions
 }
